requirements.txt
aiofiles
aiosmtplib
alembic
anyio
asgiref
async-exit-stack
async-generator
bcrypt
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
datatables
deta
dnspython
ecdsa
email-validator
fastapi
fastapi-mail
future
greenlet
gunicorn
h11
idna
importlib-metadata
Jinja2
joblib
jsondatabase
Mako
MarkupSafe
mathparse
mysql-connector-python
nltk
oauthlib
passlib
Pillow
Pint
protobuf
psycopg2
pyasn1
pycparser
pydantic
PyJWT
pymongo
python-dateutil
python-dotenv
python-editor
python-jose
python-multipart
python-twitter
regex
requests
requests-oauthlib
rsa
six
sniffio
SQLAlchemy
starlette
tqdm
typing_extensions
urllib3
uvicorn
zipp
